% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R, R/step_select_infgain.R,
%   R/step_select_mrmr.R, R/step_select_vip.R
\name{bake}
\alias{bake}
\alias{bake.step_select_infgain}
\alias{bake.step_select_mrmr}
\alias{bake.step_select_vip}
\title{Bake}
\usage{
\method{bake}{step_select_infgain}(object, new_data, ...)

\method{bake}{step_select_mrmr}(object, new_data, ...)

\method{bake}{step_select_vip}(object, new_data, ...)
}
\description{
Bake
}
\keyword{internal}
