% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_fimp.R
\name{bake.step_fimp}
\alias{bake.step_fimp}
\title{bake method to apply the method from prep to new_data}
\usage{
bake.step_fimp(object, new_data, ...)
}
\arguments{
\item{object}{is the updated step function that has been through the corresponding prep code}

\item{new_data}{is a tibble of data to be processed}

\item{...}{currently unused}
}
\description{
bake method to apply the method from prep to new_data
}
