% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_importance.R
\name{bake.step_importance}
\alias{bake.step_importance}
\title{prep method does not apply the method, it only calculates any required data.
The bake method is defined to do this.}
\usage{
bake.step_importance(object, new_data, ...)
}
\arguments{
\item{object}{is the updated step function that has been through the
corresponding prep code}

\item{new_data}{is a tibble of data to be processed}

\item{...}{currently unused}
}
\description{
prep method does not apply the method, it only calculates any required data.
The bake method is defined to do this.
}
