% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_mrmr.R
\name{bake.step_mrmr}
\alias{bake.step_mrmr}
\title{The prep method does not apply the method, it only calculates any required
data. The bake method is defined to do this.}
\usage{
bake.step_mrmr(object, new_data, ...)
}
\arguments{
\item{object}{The updated step function that has been through the
corresponding prep code.}

\item{new_data}{A tibble of data to be processed.}

\item{...}{Currently unused.}
}
\description{
The prep method does not apply the method, it only calculates any required
data. The bake method is defined to do this.
}
