% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_cubist_fs.R
\name{prep.step_cubist_fs}
\alias{prep.step_cubist_fs}
\title{Define the estimation procedure}
\usage{
prep.step_cubist_fs(x, training, info = NULL, ...)
}
\arguments{
\item{x}{the step object}

\item{training}{a tibble that has the training set data}

\item{info}{a tibble that contains information on the current set of data.
This is updated each time as each step function is evaluated by its prep
method}

\item{...}{Currently unused}
}
\description{
Define the estimation procedure
}
