% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_mrmr.R
\name{prep.step_mrmr}
\alias{prep.step_mrmr}
\title{Define the estimation procedure}
\usage{
prep.step_mrmr(x, training, info = NULL, ...)
}
\arguments{
\item{x}{The step object.}

\item{training}{A tibble that has the training set data.}

\item{info}{A tibble that contains information on the current set of data.
This is updated each time as each step function is evaluated by its prep
method.}

\item{...}{Currently unused}
}
\description{
Define the estimation procedure
}
