% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_ranger_rfe.R
\name{prep.step_ranger_rfe}
\alias{prep.step_ranger_rfe}
\title{Define the estimation procedure}
\usage{
prep.step_ranger_rfe(x, training, info = NULL, ...)
}
\arguments{
\item{x}{the step object}

\item{training}{a tibble that has the training set data}

\item{info}{a tibble that contains information on the current set of data. This is updated each
time as each step function is evaluated by its prep method}

\item{...}{Currently unused}
}
\description{
Define the estimation procedure
}
