% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_fimp.R
\name{step_fimp}
\alias{step_fimp}
\title{Select From Feature Importances (ranger)}
\usage{
step_fimp(recipe, ..., target = NULL, role = NA, trained = FALSE,
  model = NULL, threshold = NULL, to_retain = NULL, scores = NULL,
  skip = FALSE, id = rand_id("fimp"))
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of operations
for this recipe}

\item{...}{One or more selector functions to choose which variables are affected by the
step. See selections() for more details. For the tidy method, these are not currently
used}

\item{target}{character, name of response variable to use to evaluate information gain
value against the predictors}

\item{role}{Not used by this step since no new variables are created}

\item{trained}{A logical to indicate if the quantities for preprocessing have been
estimated}

\item{model}{parsnip rand_forest model specification using the `ranger` engine}

\item{threshold}{numeric, percentile of top scoring features to retain}

\item{to_retain}{character, names of features to retain}

\item{scores}{tibble, tibble of feature importance scores}

\item{skip}{A logical. Should the step be skipped when the recipe is baked by
bake.recipe()? While all operations are baked when prep.recipe() is run, some
operations may not be able to be conducted on new data (e.g. processing the outcome
variable(s)). Care should be taken when using skip = TRUE as it may affect the
computations for subsequent operations}

\item{id}{A character string that is unique to this step to identify it}
}
\value{
a step_fimp object
}
\description{
Feature selection step based on ranger feature importance scores
}
