% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_geodist_cluster.R
\name{step_geodist_cluster}
\alias{step_geodist_cluster}
\title{Information gain feature selection step}
\usage{
step_geodist_cluster(recipe, lat = NULL, lon = NULL,
  role = "predictor", num_comp = 5, trained = FALSE,
  columns = NULL, ref_lons = NULL, ref_lats = NULL, skip = FALSE,
  id = rand_id("geodist_cluster"))
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of
operations for this recipe}

\item{lat}{selector for variable to represent y coordinate data}

\item{lon}{select for variable to represent x coordinate data}

\item{role}{Not used by this step since no new variables are created}

\item{num_comp}{numeric, the number of clusters to spatially segregate the
training data into}

\item{trained}{A logical to indicate if the quantities for preprocessing have
been estimated}

\item{columns}{names for the x, y coordinates for the cluster centroids}

\item{ref_lons}{numeric, x coordinates of the cluster centroids}

\item{ref_lats}{numeric, y coordinates of the cluster centroids}

\item{skip}{A logical. Should the step be skipped when the recipe is baked by
bake.recipe()? While all operations are baked when prep.recipe() is run,
some operations may not be able to be conducted on new data (e.g.
processing the outcome variable(s)). Care should be taken when using skip =
TRUE as it may affect the computations for subsequent operations}

\item{id}{A character string that is unique to this step to identify it}
}
\value{
a step_infgain object
}
\description{
Information gain feature selection step
}
