% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_idw.R
\name{step_idw}
\alias{step_idw}
\title{Inverse distance-weighting interpolation}
\usage{
step_idw(
  recipe,
  target = NULL,
  lon = NULL,
  lat = NULL,
  role = "predictor",
  trained = FALSE,
  neighbors = NULL,
  dist_power = NULL,
  metrics = "idw",
  spatial = NULL,
  columns = NULL,
  skip = FALSE,
  id = rand_id("idw")
)
}
\arguments{
\item{recipe}{A recipe.}

\item{target}{Selector function to choose which variable will be used to
create a new feature based on the inverse distance-weighted mean of
surrounding observations.}

\item{lon, lat}{Selector functions to indicate which variables should be used
to calculate the distance measures to the closest observations.}

\item{role}{role or model term created by this step, what analysis
role should be assigned?. By default, the function assumes
that resulting distance will be used as a predictor in a model.}

\item{trained}{A logical that will be updated once the step has been trained.}

\item{neighbors}{The number of closest neighbours to use in the distance
weighting.}

\item{dist_power}{Distance weighting power. The distance-based weights are
raised to this power. Values > 1 cause very proximal observations to be
weighted more highly.}

\item{metrics}{Summary metrics to produce. One or multiple of c("idw",
"mean", "sd", "raw).}

\item{spatial}{Used internally to store an `sf` object of the spatial
locations.}

\item{columns}{Used internally to store the names of the spatial columns.}

\item{skip}{A logical to skip training.}

\item{id}{An identifier for the step. If omitted then this is generated
automatically.}
}
\value{
An updated version of `recipe` with the new step added to the
  sequence of existing steps (if any). For the `tidy` method, a tibble with
  columns echoing the values of `lat`, `lon`, `neighbors`, `dist_power`, and
  `id`.
}
\description{
`step_idw` creates a *specification* of a recipe step that will add a new
'idw' feature to a dataset based on the inverse distance-weighted mean of
surrounding observations.
}
\examples{
library(gstat)
library(sf)
library(tidymodels)

data("meuse.all")

regr <- rand_forest(mode = "regression", min_n = 1) \%>\%
    set_engine("ranger")

rec <- meuse.all \%>\%
    select(x, y, zinc, elev, dist.m, soil) \%>\%
    recipe(zinc ~ .) \%>\%
    step_idw(lon = x, lat = y, target = zinc, neighbors = 3, dist_power = 2)
}
