% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_select_carscore.R
\name{step_select_carscore}
\alias{step_select_carscore}
\alias{tidy.step_select_carscore}
\title{Information gain feature selection step}
\usage{
step_select_carscore(
  recipe,
  ...,
  outcome = NULL,
  role = NA,
  trained = FALSE,
  top_p = NA,
  threshold = NA,
  lambda = NA,
  diagonal = FALSE,
  exclude = NULL,
  scores = NULL,
  skip = FALSE,
  id = recipes::rand_id("select_carscore")
)

\method{tidy}{step_select_carscore}(x, ...)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of
operations for this recipe.}

\item{...}{One or more selector functions to choose which variables are
affected by the step. See selections() for more details. For the tidy
method, these are not currently used.}

\item{outcome}{A character string with the name of the response variable.
This must refer to a numeric feature for regression.}

\item{role}{Not used by this step since no new variables are created.}

\item{trained}{A logical to indicate if the quantities for preprocessing have
been estimated.}

\item{top_p}{An integer with the number of best scoring features to
select.}

\item{threshold}{A numeric value between 0 and 1 representing the percentile
of best scoring features to select. Features with scores that are _larger_
than the specified threshold will be retained, for example `threshold =
0.9` will retain only predictors with scores in the top 90th percentile.
Note that this overrides `top_p`.}

\item{lambda}{The correlation shrinkage intensity (range 0-1).}

\item{diagonal}{For diagonal = FALSE (the default) CAR scores are computed;
otherwise with diagonal = TRUE marginal correlations.}

\item{exclude}{A character vector of predictor names that will be removed
from the data. This will be set when `prep()` is used on the recipe and
should not be set by the user.}

\item{scores}{A tibble with 'variable' and 'scores' columns containing the
names of the variables and the absolute values of the calculated CAR
scores. This parameter is only produced after the recipe has been trained.}

\item{skip}{A logical. Should the step be skipped when the recipe is baked by
bake.recipe()? While all operations are baked when prep.recipe() is run,
some operations may not be able to be conducted on new data (e.g.
processing the outcome variable(s)). Care should be taken when using skip =
TRUE as it may affect the computations for subsequent operations.}

\item{id}{A character string that is unique to this step to identify it.}

\item{x}{A `step_select_carscore` object.}
}
\value{
A step_select_carscore object.
}
\description{
`step_select_carscore` creates a *specification* of a recipe step that
selects a subset of predictors as part of a regression model based on the
scores of the CAR score algorithm. This step requires the `care` package to be
installed. The top `top_p` scoring features, or features whose scores occur
in the top percentile `threshold` will be retained as new predictors.
}
\details{
The recipe will stop if both `top_p` and `threshold` are left unspecified.
}
\examples{
library(recipes)

data(car_prices, package = "modeldata")

rec <-
 recipe(Price ~ ., data = car_prices) \%>\%
 step_select_carscore(all_predictors(), outcome = "Price", top_p = 5, threshold = 0.7)

prepped <- prep(rec)

new_data <- juice(prepped)
prepped
}
\concept{preprocessing}
\concept{supervised_filter}
\keyword{datagen}
