% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_select_roc.R
\name{step_select_roc}
\alias{step_select_roc}
\alias{tidy.step_select_roc}
\title{Filter Numeric Predictors using ROC Curve}
\usage{
step_select_roc(
  recipe,
  ...,
  outcome,
  role = "predictor",
  trained = FALSE,
  threshold = NA,
  top_p = NA,
  exclude = NULL,
  skip = FALSE,
  id = rand_id("select_roc")
)

\method{tidy}{step_select_roc}(x, ...)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of
operations for this recipe.}

\item{...}{One or more selector functions to choose which predictors are
affected by the step. See [selections()] for more details. For the `tidy`
method, these are not currently used.}

\item{outcome}{A single character string that specifies a single categorical
variable to be used as the class.}

\item{role}{For model terms created by this step, what analysis role should
they be assigned?. By default, the function assumes that resulting distances
will be used as predictors in a model.}

\item{trained}{A logical to indicate if the quantities for preprocessing have
been estimated.}

\item{threshold}{A numeric value, in AUC units, where predictors with ROC
AUC values _larger_ than the threshold will be retained. A value of `NA`
implies that this criterion will be ignored.}

\item{top_p}{An integer that will be used to select the predictors with the
largest ROC AUC values. A value of `NA` implies that this criterion will be
ignored.}

\item{exclude}{A character vector of predictor names that will be removed
from the data. This will be set when `prep()` is used on the recipe and
should not be set by the user.}

\item{skip}{A logical. Should the step be skipped when the recipe is baked by
bake.recipe()? While all operations are baked when prep.recipe() is run,
some operations may not be able to be conducted on new data (e.g.
processing the outcome variable(s)). Care should be taken when using skip =
TRUE as it may affect the computations for subsequent operations.}

\item{id}{A character string that is unique to this step to identify it.}

\item{x}{A `step_select_roc` object.}
}
\value{
An updated version of `recipe` with the new step
 added to the sequence of existing steps (if any). For the
 `tidy` method, a tibble with a `terms` column for which predictors were
removed.
}
\description{
`step_select_roc` creates a *specification* of a recipe step that will
 filter predictors using their relationship with the outcome as measured
 using a Receiver Operating Characteristic curve.
}
\details{
The recipe will stop if both `top_p` and `threshold` are left unspecified.

The ROC AUC will be set to be 1 - AUC if the value is less than 0.50.
}
\examples{
data(cells, package = "modeldata")

rec <-
  recipe(class ~ ., data = cells[, -1]) \%>\%
  step_select_roc(all_predictors(), outcome = "class", top_p = 10, threshold = 0.9) \%>\%
  prep()

rec \%>\% juice(all_predictors()) \%>\% names()

# Use ROC values to select but always keep at least one:
rec <-
  recipe(class ~ ., data = cells[, -1]) \%>\%
  step_select_roc(all_predictors(), outcome = "class", top_p = 1, threshold = 0.99) \%>\%
  prep()

rec \%>\% juice(all_predictors()) \%>\% names()

# in case of missing data...
}
\concept{preprocessing}
\concept{supervised_filter}
\keyword{datagen}
