% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{top_p}
\alias{top_p}
\title{Parameter functions for feature selection recipes}
\usage{
top_p(range = c(1L, 4L), trans = NULL)
}
\arguments{
\item{range}{A two-element vector holding the _defaults_ for the smallest and
largest possible values, respectively.}

\item{trans}{A `trans` object from the `scales` package, such as
`scales::log10_trans()` or `scales::reciprocal_trans()`. If not provided,
the default is used which matches the units used in `range`. If no
transformation, `NULL`.}
}
\value{
A function with classes "quant_param" and "param"
}
\description{
Feature selection recipes allow the top-performing features to be selected
using two parameters. `top_p` is for specifying the number of the
top-performing features.
}
\examples{
top_p(c(3, 10))
}
