% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_select_infgain.R, R/step_select_mrmr.R,
%   R/step_select_roc.R, R/step_select_vip.R, R/step_select_xtab.R, R/tunable.R
\name{tunable.step_select_infgain}
\alias{tunable.step_select_infgain}
\alias{tunable.step_select_mrmr}
\alias{tunable.step_select_roc}
\alias{tunable.step_select_vip}
\alias{tunable.step_select_xtab}
\alias{tunable}
\title{Determines which parameters in an object can be used along with information
about the parameters}
\usage{
tunable.step_select_infgain(x, ...)

tunable.step_select_mrmr(x, ...)

tunable.step_select_roc(x, ...)

tunable.step_select_vip(x, ...)

tunable.step_select_xtab(x, ...)
}
\arguments{
\item{x}{An object, such as a recipe, recipe step, or parsnip model
specification.}

\item{...}{No currently used.}
}
\description{
Determines which parameters in an object can be used along with information
about the parameters
}
