% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_infgain.R
\name{tunable.step_infgain}
\alias{tunable.step_infgain}
\title{Specify tunable arguments of step}
\usage{
tunable.step_infgain(x, ...)
}
\arguments{
\item{x}{step}

\item{...}{currently unused}
}
\value{
tibble
}
\description{
Specify tunable arguments of step
}
