% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_mrmr.R
\name{tunable.step_mrmr}
\alias{tunable.step_mrmr}
\title{Specify tunable arguments of step}
\usage{
tunable.step_mrmr(x, ...)
}
\arguments{
\item{x}{step}

\item{...}{currently unused}
}
\value{
tibble
}
\description{
Specify tunable arguments of step
}
