% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_ranger_fs.R
\name{tunable.step_ranger_fs}
\alias{tunable.step_ranger_fs}
\title{Specify tunable arguments of step}
\usage{
tunable.step_ranger_fs(x, ...)
}
\arguments{
\item{x}{step}

\item{...}{currently unused}
}
\value{
tibble
}
\description{
Specify tunable arguments of step
}
