% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_importances.R
\name{pull_importances}
\alias{pull_importances}
\title{Pull feature importances from a `model_fit` object}
\usage{
pull_importances(object, ...)
}
\arguments{
\item{object}{A `model_fit` object.}

\item{...}{A list of other parameters passed to the feature importance
method.}
}
\value{
tibble
}
\description{
Pull feature importances from a `model_fit` object
}
\examples{
\dontrun{
model <- boost_tree(mode = "classification") \%>\%
set_engine("xgboost")
model_fit <- model \%>\% fit(Species ~., iris)
pull_importances(model_fit)
}
}
