% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_mrmr.R
\name{step_mrmr}
\alias{step_mrmr}
\title{mRMR feature selection recipe}
\usage{
step_mrmr(recipe, ..., role = NA, trained = FALSE, k = NULL,
  threads = 0, to_retain = NULL, skip = FALSE,
  id = rand_id("mrmr"))
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of operations
for this recipe}

\item{...}{One or more selector functions to choose which variables are affected by the
step. See selections() for more details. For the tidy method, these are not currently
used}

\item{role}{Not used by this step since no new variables are created}

\item{trained}{A logical to indicate if the quantities for preprocessing have been
estimated}

\item{k}{integer, number of best features to select}

\item{threads}{integer, number of threads to use for processing, default = 0 uses all
available threads}

\item{to_retain}{character, names of features to retain}

\item{skip}{A logical. Should the step be skipped when the recipe is baked by
bake.recipe()? While all operations are baked when prep.recipe() is run, some
operations may not be able to be conducted on new data (e.g. processing the outcome
variable(s)). Care should be taken when using skip = TRUE as it may affect the
computations for subsequent operations}

\item{id}{A character string that is unique to this step to identify it}
}
\value{
a step_mrmr object
}
\description{
Initial function - simple wrapper around add_step
}
